Serbian Letters ALMOST done
Problem with Full Name vs First Name or Second  name FIX. There was the problem because comma or apostrophe c an be part of the full name but it can't be part of the first name. Second name can contain apostrophe but it can't contain , The SERBIAN LETTERS are done except in subscribe and there are couple more small flaws.